# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Quran Qaida Class Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List all students (teacher only)
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudentWithUser"
    post:
      operationId: createStudent
      tags: [students]
      summary: Create a new student profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudentBody"
      responses:
        "201":
          description: Created student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
...
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get a student by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentWithUser"
    put:
      operationId: updateStudent
      tags: [students]
      summary: Update a student
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStudentBody"
      responses:
        "200":
          description: Updated student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    delete:
      operationId: deleteStudent
      tags: [students]
      summary: Delete a student
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /topics:
    get:
      operationId: listTopics
      tags: [topics]
      summary: List all topics
      responses:
        "200":
          description: List of topics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Topic"
    post:
      operationId: createTopic
      tags: [topics]
      summary: Create a new topic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTopicBody"
      responses:
        "201":
          description: Created topic
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Topic"
  /topics/{id}:
    get:
      operationId: getTopic
      tags: [topics]
      summary: Get a topic with its questions
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Topic with questions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TopicWithQuestions"
    put:
      operationId: updateTopic
      tags: [topics]
      summary: Update a topic
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTopicBody"
      responses:
        "200":
          description: Updated topic
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Topic"
    delete:
      operationId: deleteTopic
      tags: [topics]
      summary: Delete a topic
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          nullable: true
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        profileImageUrl:
          type: string
          nullable: true
        role:
          type: string
          nullable: true
      required:
        - id
    Student:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        scheduleType:
          type: string
        currentLesson:
          type: integer
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - scheduleType
        - currentLesson
    StudentWithUser:
      allOf:
        - $ref: "#/components/schemas/Student"
        - type: object
          properties:
            user:
              $ref: "#/components/schemas/UserProfile"
    CreateStudentBody:
      type: object
      properties:
        userId:
          type: string
        scheduleType:
          type: string
        currentLesson:
          type: integer
        notes:
          type: string
      required:
        - userId
        - scheduleType
    UpdateStudentBody:
      type: object
      properties:
        scheduleType:
          type: string
        currentLesson:
          type: integer
        notes:
          type: string
    Topic:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
    TopicWithQuestions:
      allOf:
        - $ref: "#/components/schemas/Topic"
        - type: object
          properties:
            questions:
              type: array
              items:
                $ref: "#/components/schemas/Question"
    CreateTopicBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
      required:
        - title
    Question:
      type: object
      properties:
        id:
          type: integer
        topicId:
          type: integer
        question:
          type: string
        questionType:
          type: string
        options:
          type: array
          items:
            type: string
          nullable: true
        correctAnswer:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - topicId
        - question
        - questionType
        - correctAnswer
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...