# lib/api-zod/src/generated/types/studentWithUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quran Qaida Class Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { Student } from "./student";
import type { UserProfile } from "./userProfile";

export type StudentWithUser = Student & {
  user?: UserProfile;
};