# lib/api-zod/src/generated/types/updateStudentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quran Qaida Class Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateStudentBody {
  scheduleType?: string;
  currentLesson?: number;
  notes?: string;
}